+++ Changes.md
@@ -7,8 +7,8 @@ Version 1.10.0
 
 * Source code moved to GitHub [ij-plugins/ijp-toolkit](https://github.com/ij-plugins/ijp-toolkit)
 * New [SBT](http://www.scala-sbt.org/)-based build system and new packaging approach.
-* Coherence Enhancing Diffusion retired here, it was moved and significantly
-  improved in [ij-plugins/ijp-toolkit2](https://github.com/ij-plugins/ijp-toolkit2).
+* Coherence Enhancing Diffusion was rewritten and significantly improved.
+  The new algorithm is over 7 times faster (on a quad-core processor).
 * Seeded Region Growing Tool: add button to create seed image.
 * MetaImage Reader/Writer respect `Origin` tag, mapped into calibration x.y.z origin.
 * MetaImage Reader - does not break on unknown tags, just ignores them.